back/requirements.txt
uvicorn
fastapi
requests
alembic
psycopg2-binary